[SP-191] tweak coverage to omit the tests that not need testing
