+++ app/Validators/Requests/Doctor/VaccinationSlot/FilterVaccinationSlotValidator.cs
@@ -1,8 +1,10 @@
-﻿using backend.Dto.Requests.Doctor.VaccinationSlot;
+﻿using System.Diagnostics.CodeAnalysis;
+using backend.Dto.Requests.Doctor.VaccinationSlot;
 using FluentValidation;
 
 namespace backend.Validators.Requests.Doctor.VaccinationSlot
 {
+    [ExcludeFromCodeCoverage]
     public class FilterVaccinationValidator : AbstractValidator<FilterVaccinationSlotsRequest>
     {
         public FilterVaccinationValidator()
