+++ app/Validators/Requests/Patient/FilterVaccinationsValidator.cs
@@ -1,8 +1,10 @@
-﻿using backend.Dto.Requests.Patient;
+﻿using System.Diagnostics.CodeAnalysis;
+using backend.Dto.Requests.Patient;
 using FluentValidation;
 
 namespace backend.Validators.Requests.Patient
 {
+    [ExcludeFromCodeCoverage]
     public class FilterVaccinationsValidator : AbstractValidator<FilterVaccinationsRequest>
     {
         public FilterVaccinationsValidator()
